requirements.txt
alembic==1.13.2
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
asyncpg==0.29.0
bcrypt==4.2.0
billiard==4.2.1
black==24.8.0
celery==5.4.0
certifi==2024.7.4
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.111.1
fastapi-cache2==0.2.2
fastapi-cli==0.0.4
greenlet==3.1.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
Jinja2==3.1.4
kombu==5.4.2
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.1
passlib==1.7.4
pathspec==0.12.1
pendulum==3.0.0
pillow==11.0.0
platformdirs==4.3.6
pluggy==1.5.0
prompt_toolkit==3.0.48
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.9.0
pyright==1.1.388
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-dotenv==0.5.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
redis==5.1.1
rich==13.7.1
ruff==0.7.3
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.37.2
time-machine==2.16.0
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.30.3
uvloop==0.19.0
vine==5.1.0
watchfiles==0.22.0
wcwidth==0.2.13
websockets==12.0
